feat(settings): merge Wallet + Spend + Usage into one Billing panel
The three money panels each rendered their own balance/session card — the
same thing three times — and split spend routing, the ledger, and usage
analytics across separate tabs. Collapse them into a single Billing panel
with four sections: Balance (the one canonical card) · Spend routing (LLM +
Compute managed/BYOK) · Usage (weekly bars, per-model breakdown, editable
budget) · Recent transactions. Each concern loads from one real endpoint
independently (wallet / billing / usage+prefs) with its own error state, so
a failure in one never blanks the others; balance + usage refetch on focus.

Every control stays wired — no dead buttons. Removes Wallet.tsx, Spend.tsx,
Usage.tsx. Settings workspace section is now Permissions · Sandbox ·
Credentials · Billing · Storage · General.
