+++ frontend/workspace/src/components/settings/registry.ts
@@ -32,10 +32,8 @@ export type SettingsPanelId =
   | "permissions"
   | "sandbox"
   | "credentials"
-  | "spend"
-  | "wallet"
+  | "billing"
   | "storage"
-  | "usage"
   | "general"
 
 export interface SettingsPanel {
@@ -120,10 +118,10 @@ export const SETTINGS_PANELS: SettingsPanel[] = [
     section: "workspace",
     component: lazy(() => import("./Credentials")),
   },
-  { id: "spend", title: "Spend", icon: "sliders", section: "workspace", component: lazy(() => import("./Spend")) },
-  { id: "wallet", title: "Wallet", icon: "checklist", section: "workspace", component: lazy(() => import("./Wallet")) },
+  // Wallet + Spend + Usage merged into one Billing panel (they each rendered a
+  // duplicate balance card). Balance · Spend routing · Usage · Ledger.
+  { id: "billing", title: "Billing", icon: "sliders", section: "workspace", component: lazy(() => import("./Billing")) },
   { id: "storage", title: "Storage", icon: "folder", section: "workspace", component: lazy(() => import("./Storage")) },
-  { id: "usage", title: "Usage", icon: "bullet-list", section: "workspace", component: lazy(() => import("./Usage")) },
   {
     id: "general",
     title: "General",
